requirements.txt
asgiref==3.7.2
configuration-settings==1.0.0
confusable-homoglyphs==3.2.0
crispy-bootstrap5==2023.10
dj-database-url==2.1.0
Django==4.2.7
django-autoslug==1.9.9
django-configuration==0.0.3
django-configurations==2.5
django-crispy-forms==2.1
django-etc==1.4.0
django-hitcount==1.3.5
django-jazzmin==2.6.0
django-polymorphic==3.1.0
django-registration==3.4
django-registration-bootstrap==0.2.0
django-registration-email==0.10
nodejs==0.1.1
optional-django==0.1.0
Pillow==10.1.0
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
